build: buildinfo: move buildinfo struct in internal api
use the buildinfo.json data for `pkg/version` instead
of injecting the values at link time.
This simplifies the builds and enables better introspection
capabilities.
