+++ .gitignore
@@ -25,3 +25,5 @@ testbin/*
 *~
 
 coverage.out
+
+pkg/version/_buildinfo.json
